--- src/widgets/calendar/edit-schedule/index.ts
@@ -1,1 +1,0 @@
-export { useEditScheduleForm } from './hooks';
